+++ Merge_Sort.h
@@ -8,9 +8,9 @@
 using namespace std;
 using namespace chrono;
 
-void mergeComparision(int *arr, int first, int mid, int last, int &comparision);
-void mergeSortComparision(int *arr, int first, int last, int &comparision);
-void calComparisionMergeSort(int *arr, int n, int &comparision);
+void mergeComparision(int *arr, int first, int mid, int last, long long &comparision);
+void mergeSortComparision(int *arr, int first, int last, long long &comparision);
+void calComparisionMergeSort(int *arr, int n, long long &comparision);
 
 void mergeTime(int *arr, int first, int mid, int last);
 void mergeSortTime(int *arr, int first, int last);
