+++ ShellSort.h
@@ -7,5 +7,5 @@ using namespace chrono;
 
 double ShellSortRuntime(int* arr, int n);
 
-void Com_ShellSort(int* arr, int n, int& cnt_com);
+void Com_ShellSort(int* arr, int n, long long& cnt_com);
