+++ bubbleSort.cpp
@@ -1,6 +1,6 @@
 #include "bubbleSort.h"
 
-void bubbleSort(int *arr, int n, int &count_comparison){
+void bubbleSort(int *arr, int n, long long &count_comparison){
     count_comparison = 0;
     for(int i = 0; ++count_comparison && i < n - 1; i++){
         for(int j = 0; ++count_comparison && j < n - i - 1; j++){
